fix(core): modifying encryption script and test to pass
summary: utilising ethPrivKey to test seeding of encryption for deterministic key generation. NOTE WARNING utilising this in production is ill advised, as two like signed messages may be used to reverse engineer the privKey. The deterministic nature is useful, however, as it adds a layer of recoverability. A prng may be satisfactory.