fix(core): removed the reliance on eth key as seed in ts implementation of rsa key generation
summer: alike to recent commits, this is the effective change, rather utilising a prng as a seed to create an rsa keypair for testing., In realiy, better methods exist for determinstic key generation, see https://datatracker.ietf.org/doc/html/rfc6979, which may be implemented in the future. It is a safety precaution as a quick fix, this suffices before rsa keys are published on chain in testing environments.